Restructure creation, perms for bind mounted dirs
Creating /openstack in a task targetted at the host, prevents a race
when it is created in a task targetted at containers but delegated to
the host. Same problem also exists with common subdirs.

/var/log/nova has a similar problem, for metal 'containers' it should be
a symlink to /openstack/log/container_name, but for lxc containers
/var/log/nova should be an empty dir ready for logs.

The log dirs have their group owner set to adm. The idea is to use a
group that exists by default in ubuntu so it will have a consistent GID,
for sharing logs between containers and hosts via bind mounts. This
group can then be given read access while each service user has
ownership and write access. Where packages interfere with patch rights
the rights have been set to allow all users rx rights to the
directories.

Change-Id: I6366535a76562e9f95cc1e0352daf76d9ae39e3b
Closes-Bug: #1407121
Related-Bug: #1403890