[wip] Add role for tempest to allow set image location
Recent glance policy additions prevent non admins from setting image
location and therefore creating remote images. Tempest users need this
functionality, so this patch adds a role to those users and a policy to
allow that role to set image location.

The change that allows users to specify roles to be added to tempest
users has already been merged upstream:
https://review.openstack.org/#/c/147542/

This patch also bumps the tempest version to latest, and
python-glanceclient to 0.15.0.

Change-Id: I9ed053bf70680db54b1cf679a171d61ea639f19d
Closes-Bug: #1411599