Read affinity from environment
This patch allows a user to set the number of each type of container in
their environment before running the aio scripts. Bootstrap-aio will
then substitute these values into the user_config.

This is the beginning
of a process for splitting the current AIO into multiple gate jobs.

Change-Id: If90704a50414150ca37584e1f823ba0e85e287ac
Partial-Blueprint: split-aio-gates