Build wheels for Non-PyPi modules
The ceph client libraries are not yet available on pypi, so it is
difficult to deploy them into venvs. Currently symlinks are created to
the host's site packages dir where the modules have been installed by
operating system packages.

This patch adds a mechanism for building wheels from modules that don't
have packaging boilerplate. The user specifies package meta data
and URLs to the relevant module, this information is used to template
setup.cfg and build wheels, these wheels are then added to the relevant
venvs.

Change-Id: Ib077e288e8c8986248d860ff36a743164b477568